Character creation updates
Creation page is now a new window that opens, and the tab on the tabbed pane was replaced with some buttons. Now there is a submit button that writes the selections to a newly added DB, this will store all PC Character information eventually. Player Character object created, not yet in use.