Ability score boxes
Created a new extension of JPanel that will hold the ability scores and their modifiers. it has its own custom border that is defined using a class that implements Border. Frame resizes better when switching pages in creation window. Submit button now makes a character object before storing it in the data base.