Skill box is made better
The skill box class now uses a skill object to get its data, rather than the enum it was using before. This object is created at the stat page level, and pulls the list from a new table on the ogl.db. A new functions was added to Data Access for this. The ability score and skill classes are now in use. The PC class now makes ability score objects.