More responsive scores and skills
A New class was added that defines a change listener. It uses the underlying document of a text box and notices when any change is made so the user no longer needs to hit enter after input in the score boxes. To account for this change, the submit button now checks all the scores to see if they are entered, rather than saving them as they are entered since the user might not hit enter at all.